Multiply emulation speed by 2 - it was running way too slow before. Need to look into this in future as I don't think the timing is completely correct

+++ src/include/MasterSystem.h
@@ -16,7 +16,7 @@ public:
 
     bool init(std::string romFilename) final;
 
-    unsigned int tick();
+    double tick();
 
     bool isRunning() final;
